Move custom attribute condition evaluation logic to its own module. Make condition evaluator generic to any kind of leaf evaulator. Update audience evaluator to use the generic condition evaluator, providing a leaf evaluator that uses the custom attribute condition evaluator module.
